backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
defusedxml==0.7.1
deptry==0.23.0
distlib==0.3.9
Django==5.1.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-filter-stubs==0.1.3
django-rosetta==0.10.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-stubs==3.14.5
djangorestframework_simplejwt==5.4.0
djoser==2.3.0
drf-extra-fields==3.7.0
dulwich==0.22.8
exceptiongroup==1.2.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
importlib_metadata==8.7.0
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
joblib==1.3.2
keyring==25.6.0
more-itertools==10.7.0
msgpack==1.1.1
mypy==1.9.0
mypy-extensions==1.0.0
nltk==3.9
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pbs-installer==2025.6.12
pillow==10.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
polib==1.2.0
psycopg2==2.9.9
pycparser==2.21
Pygments==2.19.2
PyJWT==2.8.0
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
RapidFuzz==3.13.0
regex==2023.12.25
reportlab==4.1.0
reportlab-stubs==3.6.9.post0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.13.0
ruff==0.3.7
scipy==1.12.0
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.3
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.13.3
tqdm==4.66.3
trove-classifiers==2025.5.9.12
types-psycopg2==2.9.21.20240311
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
types-regex==2023.12.25.20240311
types-requests==2.31.0.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.31.2
webcolors==1.13
webcolors-stubs==0.0.3
zipp==3.23.0
zstandard==0.23.0
